Move utilities to utils.py
