+++ tests/tpch/conftest.py
@@ -10,7 +10,7 @@ import filelock
 import pytest
 from dask.distributed import LocalCluster, performance_report
 
-from .utils import get_dataset_path
+from .utils import get_cluster_spec, get_dataset_path, get_single_vm_spec
 
 ##################
 # Global Options #
@@ -139,36 +139,7 @@ def benchmark_time(test_run_benchmark, module, scale, name):
 
 @pytest.fixture(scope="session")
 def cluster_spec(scale):
-    everywhere = dict(
-        idle_timeout="1h",
-        wait_for_workers=True,
-        scheduler_vm_types=["m6i.xlarge"],
-    )
-    if scale == 10:
-        return {
-            "worker_vm_types": ["m6i.large"],
-            "n_workers": 8,
-            **everywhere,
-        }
-    elif scale == 100:
-        return {
-            "worker_vm_types": ["m6i.large"],
-            "n_workers": 16,
-            **everywhere,
-        }
-    elif scale == 1000:
-        return {
-            "worker_vm_types": ["m6i.xlarge"],
-            "n_workers": 32,
-            **everywhere,
-        }
-    elif scale == 10000:
-        return {
-            "worker_vm_types": ["m6i.xlarge"],
-            "n_workers": 32,
-            "worker_disk_size": 200,
-            **everywhere,
-        }
+    return get_cluster_spec(scale)
 
 
 @pytest.fixture(scope="module")
@@ -290,23 +261,7 @@ def fs(local):
 
 @pytest.fixture(scope="session")
 def machine_spec(scale):
-    if scale == 10:
-        return {
-            "vm_type": "m6i.4xlarge",
-        }
-    elif scale == 100:
-        return {
-            "vm_type": "m6i.8xlarge",
-        }
-    elif scale == 1000:
-        return {
-            "vm_type": "m6i.32xlarge",
-        }
-    elif scale == 10000:
-        return {
-            "vm_type": "m6i.32xlarge",
-            "disk_size": 1000,
-        }
+    return get_single_vm_spec(scale)
 
 
 @pytest.fixture(scope="module")
