+++ tests/tpch/generate_data.py
@@ -8,7 +8,6 @@ import warnings
 from collections import namedtuple
 from typing import Union
 
-import boto3
 import botocore.session
 import click
 import coiled
@@ -18,6 +17,7 @@ import psutil
 import pyarrow as pa
 import pyarrow.compute as pc
 import pyarrow.parquet as pq
+from utils import get_bucket_region
 
 REGION = None
 SMALL_TABLES = {"region", "nation"}
@@ -301,16 +301,6 @@ def _alter_tables(con):
             "column_name"
         ):
             con.sql(f"alter table {table} alter {column} type timestamp_s")
-
-
-def get_bucket_region(path: str):
-    if not path.startswith("s3://"):
-        raise ValueError(f"'{path}' is not an S3 path")
-    bucket = path.replace("s3://", "").split("/")[0]
-    resp = boto3.client("s3").get_bucket_location(Bucket=bucket)
-    # Buckets in region 'us-east-1' results in None, b/c why not.
-    # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3/client/get_bucket_location.html#S3.Client.get_bucket_location
-    return resp["LocationConstraint"] or "us-east-1"
 
 
 @click.command()
